Fixed logic
Fixed in this commit:
1) Can't block city with active buildings
2) Started fixing the same with buildings and rooms
3) Updated database: added logical parameter "active" for Room entity

+++ src/main/java/ru/proba/DTO/building/BuildingIdDTO.java
@@ -10,5 +10,5 @@ import lombok.Setter;
 @AllArgsConstructor
 public class BuildingIdDTO {
     @NotNull
-    private Integer buildingId;
+    private Long buildingId;
 }
